// lib/api/schemas/systemType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Eigora API
 * Physics simulation backend for the Eigora platform.
 * OpenAPI spec version: 0.2.2
 */

export type SystemType = typeof SystemType[keyof typeof SystemType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SystemType = {
  two_level: 'two_level',
  harmonic: 'harmonic',
  spin: 'spin',
  rotor: 'rotor',
  ideal_gas: 'ideal_gas',
} as const;