// lib/api/schemas/thermodynamicsRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Eigora API
 * Physics simulation backend for the Eigora platform.
 * OpenAPI spec version: 0.2.2
 */
import type { SystemSchema } from './systemSchema';
import type { EnsembleSchema } from './ensembleSchema';
import type { TemperatureRange } from './temperatureRange';

export interface ThermodynamicsRequest {
  system?: SystemSchema;
  ensemble?: EnsembleSchema;
  temperatures?: TemperatureRange;
}